Updated fight system and reformatted code
Fight system now has proper attacking. It clones the player instance meaning your health is reset after every fight. Eventually I aim for fights to give rewards or maybe xp and to possibly have a system where enemies can get more powerful over time. I'll need to make it so the type of enemy, it's stats and if it's unique somehow is computed and passed in when encountering enemies